updated tests to work with new HTC-related script changes
